refactor: snapshot module for deno_runtime
