+++ runtime/lib.rs
@@ -30,6 +30,7 @@ pub mod inspector_server;
 pub mod js;
 pub mod ops;
 pub mod permissions;
+pub mod snapshot;
 pub mod tokio_util;
 pub mod web_worker;
 pub mod worker;
